requirements.txt
flask==2.3.3
werkzeug==2.3.7
PyPDF2==3.0.1
PyMuPDF==1.23.22
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.0
openai==1.3.0
python-dotenv==1.0.0 